update to m81

+++ src/opts/SkBitmapProcState_opts.h
@@ -400,8 +400,8 @@ static void filter_and_scale_by_alpha(
     tmp = vmul_u16(tmp, vscale);  // multiply result by scale
   }
 
-  vres = vshrn_n_u16(vcombine_u16(tmp, vcreate_u16(0)), 8);  // shift down result by 8
-  vst1_lane_u32(dst, vreinterpret_u32_u8(vres), 0);          // store result
+  vres = vshrn_n_u16(vcombine_u16(tmp, vcreate_u16((uint64_t)0)), 8);  // shift down result by 8
+  vst1_lane_u32(dst, vreinterpret_u32_u8(vres), 0);                    // store result
 }
 #  else
 static void filter_and_scale_by_alpha(
